fixing task handler issues; refactoring
