+++ src/galileo/utils/core_api_client.py
@@ -140,6 +140,8 @@ class GalileoCoreApiClient:
 
         json = traces_ingest_request.model_dump(mode="json")
 
+        print(f"calling ingest traces for {traces_ingest_request.traces[0].id}")
+
         return await self._make_async_request(
             RequestMethod.POST, endpoint=Routes.traces.format(project_id=self.project_id), json=json
         )
@@ -157,13 +159,25 @@ class GalileoCoreApiClient:
         )
 
     async def ingest_spans(self, spans_ingest_request: SpansIngestRequest) -> dict[str, str]:
+        if self.experiment_id:
+            spans_ingest_request.experiment_id = UUID(self.experiment_id)
+        elif self.log_stream_id:
+            spans_ingest_request.log_stream_id = UUID(self.log_stream_id)
+
         json = spans_ingest_request.model_dump(mode="json")
+
+        print(f"calling ingest spans for {spans_ingest_request.spans[0].id}")
 
         return await self._make_async_request(
             RequestMethod.POST, endpoint=Routes.spans.format(project_id=self.project_id), json=json
         )
 
     def ingest_spans_sync(self, spans_ingest_request: SpansIngestRequest) -> dict[str, str]:
+        if self.experiment_id:
+            spans_ingest_request.experiment_id = UUID(self.experiment_id)
+        elif self.log_stream_id:
+            spans_ingest_request.log_stream_id = UUID(self.log_stream_id)
+
         json = spans_ingest_request.model_dump(mode="json")
 
         return self._make_request(
@@ -177,6 +191,8 @@ class GalileoCoreApiClient:
             trace_update_request.log_stream_id = UUID(self.log_stream_id)
 
         json = trace_update_request.model_dump(mode="json")
+
+        print(f"calling update trace for {trace_update_request.trace_id}")
 
         return await self._make_async_request(
             RequestMethod.PATCH,
@@ -205,6 +221,8 @@ class GalileoCoreApiClient:
             span_update_request.log_stream_id = UUID(self.log_stream_id)
 
         json = span_update_request.model_dump(mode="json")
+
+        print(f"calling update span for {span_update_request.span_id}")
 
         return await self._make_async_request(
             RequestMethod.PATCH,
